Register EBO in VAO
And make the code structure clearer.

I think what I had previously was not the way you're intended to do
this. If I understand this code, I'm not storing (references to) *both*
the VBO and EBO in the VAO now.

To improve readability, I now bind the VAO first, and also explicitly
unbind everything at the end of initialization.