Make object rotate around its own axis
And zoom in camera.

The changes to mat4f_perspective are just to change the zoom level.

We accomplish the axis rotation by moving the cube to have its xy-center
to be at (0, 0), and then just rotating around the z-axis at (0, 0).

Then we need to translate the box into the part of the scene where it is
actually visible, which we accomplish with the new mat4f_translation.